fix(ci): satisfy session settings quality gates
