+++ apps/.i18n/native/ar.json
@@ -1074,6 +1074,11 @@
       "translated": "تعذّر تحديث النموذج."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "تعذّر تحديث مستوى التفكير."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "فشلت المحادثة قبل بدء التشغيل؛ حاول مرة أخرى."
