+++ apps/.i18n/native/de.json
@@ -1074,6 +1074,11 @@
       "translated": "Das Modell konnte nicht aktualisiert werden."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "Die Denkstufe konnte nicht aktualisiert werden."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "Der Chat ist vor Beginn der Ausführung fehlgeschlagen. Versuchen Sie es erneut."
