+++ apps/.i18n/native/es.json
@@ -1074,6 +1074,11 @@
       "translated": "No se pudo actualizar el modelo."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "No se pudo actualizar el nivel de razonamiento."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "El chat falló antes de que se iniciara la ejecución; inténtalo de nuevo."
