+++ apps/.i18n/native/fa.json
@@ -1074,6 +1074,11 @@
       "translated": "مدل به‌روزرسانی نشد."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "سطح تفکر به‌روزرسانی نشد."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "گفت‌وگو پیش از شروع اجرا ناموفق بود؛ دوباره تلاش کنید."
