+++ apps/.i18n/native/it.json
@@ -1074,6 +1074,11 @@
       "translated": "Impossibile aggiornare il modello."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "Impossibile aggiornare il livello di ragionamento."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "La chat non è riuscita prima dell'avvio dell'esecuzione; riprova."
