+++ apps/.i18n/native/ja-JP.json
@@ -1074,6 +1074,11 @@
       "translated": "モデルを更新できませんでした。"
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "思考レベルを更新できませんでした。"
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "実行開始前にチャットが失敗しました。もう一度お試しください。"
