+++ apps/.i18n/native/nl.json
@@ -1074,6 +1074,11 @@
       "translated": "Kan het model niet bijwerken."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "Kon het denkniveau niet bijwerken."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "De chat is mislukt voordat de uitvoering begon; probeer het opnieuw."
