+++ apps/.i18n/native/pt-BR.json
@@ -1074,6 +1074,11 @@
       "translated": "Não foi possível atualizar o modelo."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "Não foi possível atualizar o nível de raciocínio."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "O chat falhou antes do início da execução; tente novamente."
