+++ apps/.i18n/native/ru.json
@@ -1074,6 +1074,11 @@
       "translated": "Не удалось обновить модель."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "Не удалось обновить уровень рассуждений."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "Сбой чата до начала выполнения; попробуйте ещё раз."
