+++ apps/.i18n/native/th.json
@@ -1074,6 +1074,11 @@
       "translated": "ไม่สามารถอัปเดตโมเดลได้"
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "ไม่สามารถอัปเดตระดับการคิดได้"
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "แชตล้มเหลวก่อนเริ่มการทำงาน โปรดลองอีกครั้ง"
