+++ apps/.i18n/native/tr.json
@@ -1074,6 +1074,11 @@
       "translated": "Model güncellenemedi."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "Düşünme düzeyi güncellenemedi."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "Sohbet, çalıştırma başlamadan önce başarısız oldu; tekrar deneyin."
