+++ apps/.i18n/native/uk.json
@@ -1074,6 +1074,11 @@
       "translated": "Не вдалося оновити модель."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "Не вдалося оновити рівень міркування."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "Не вдалося розпочати чат; спробуйте ще раз."
