+++ apps/.i18n/native/vi.json
@@ -1074,6 +1074,11 @@
       "translated": "Không thể cập nhật mô hình."
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "Không thể cập nhật mức độ suy luận."
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "Cuộc trò chuyện gặp lỗi trước khi bắt đầu chạy; hãy thử lại."
