+++ apps/.i18n/native/zh-CN.json
@@ -1074,6 +1074,11 @@
       "translated": "无法更新模型。"
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "无法更新思考级别。"
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "聊天在运行开始前失败；请重试。"
