+++ apps/.i18n/native/zh-TW.json
@@ -1074,6 +1074,11 @@
       "translated": "無法更新模型。"
     },
     {
+      "id": "native.android.e4ab8aaa795ba79f",
+      "source": "Could not update thinking level.",
+      "translated": "無法更新思考層級。"
+    },
+    {
       "id": "native.android.956a394937480f8a",
       "source": "Chat failed before the run started; try again.",
       "translated": "聊天在執行開始前失敗；請再試一次。"
